+++ kamehouse-shell/bin/common/bashrc/bashrc.sh
@@ -5,6 +5,6 @@ source ${HOME}/programs/kamehouse-shell/bin/common/bashrc/prompt.sh
 source ${HOME}/programs/kamehouse-shell/bin/common/bashrc/message.sh
 source ${HOME}/programs/kamehouse-shell/bin/common/bashrc/path.sh
 source ${HOME}/programs/kamehouse-shell/bin/common/bashrc/alias.sh
-source ${HOME}/programs/kamehouse-shell/bin/kamehouse/set-java-home.sh false
+source ${HOME}/programs/kamehouse-shell/bin/kamehouse/set-java-home.sh false false
 
 export EDITOR=vim