+++ kamehouse-shell/bin/kamehouse/set-java-home.sh
@@ -1,10 +1,22 @@
-# call this script with 'source PATH-TO-SCRIPT/set-java-home.sh false' to skip logging
-# call this script with 'source PATH-TO-SCRIPT/set-java-home.sh true' to log info
+# param1: [true|false] override existing definition of JAVA_HOME
+# param2: [true|false] log info messages
+# call this script with 'source PATH-TO-SCRIPT/set-java-home.sh true false' to skip logging
+# call this script with 'source PATH-TO-SCRIPT/set-java-home.sh true true' to log info
 
 main() {
-  LOG_INFO=$1
-  logger.info "Setting JAVA_HOME" "${LOG_INFO}"
+  OVERRIDE=$1
+  LOG_INFO=$2
+  logger.info "Current JAVA_HOME=${JAVA_HOME} OVERRIDE=${OVERRIDE}"  "${LOG_INFO}"
+  if [[ "${OVERRIDE}" == "false" && -n "${JAVA_HOME}" ]]; then
+    logger.info "Using already set JAVA_HOME=${JAVA_HOME}" "${LOG_INFO}"
+    return
+  fi
+  setJavaHome "${LOG_INFO}"
+}
 
+setJavaHome() {
+  LOG_INFO=$1
+  logger.info "Setting new value for JAVA_HOME" "${LOG_INFO}"
   ### Java 17
   # Sort them in order from less priority to highest priority 
   # so the last one that matches is the one I want
@@ -17,7 +29,6 @@ main() {
   if [ -d "/usr/lib/jvm/java-17-openjdk-amd64" ]; then
     export JAVA_HOME=/usr/lib/jvm/java-17-openjdk-amd64
   fi
-
   logger.info "JAVA_HOME=${JAVA_HOME}" "${LOG_INFO}"
 }